+++ internal/application/usecases/get_stock.go
@@ -11,7 +11,7 @@ type GetStock struct {
 	sr domain.StockRepository
 }
 
-func (uc *GetStock) Execute(ctx context.Context, stockID uint, userID *uint) (*domain.PopulatedStock, error) {
+func (uc *GetStock) Execute(ctx context.Context, stockID uint64, userID *uint64) (*domain.PopulatedStock, error) {
 	stock, err := uc.sr.Get(ctx, stockID, userID)
 	if err != nil {
 		return nil, err
